+++ JAG3D/src/org/applied_geodesy/jag3d/ui/JAG3D.java
@@ -44,7 +44,6 @@ import org.applied_geodesy.jag3d.ui.dialog.CongruentPointDialog;
 import org.applied_geodesy.jag3d.ui.dialog.FormatterOptionDialog;
 import org.applied_geodesy.jag3d.ui.dialog.LeastSquaresSettingDialog;
 import org.applied_geodesy.jag3d.ui.dialog.NetworkAdjustmentDialog;
-import org.applied_geodesy.jag3d.ui.dialog.OptionDialog;
 import org.applied_geodesy.jag3d.ui.dialog.ProjectionAndReductionDialog;
 import org.applied_geodesy.jag3d.ui.dialog.RankDefectDialog;
 import org.applied_geodesy.jag3d.ui.dialog.SearchAndReplaceDialog;
@@ -56,6 +55,7 @@ import org.applied_geodesy.jag3d.ui.menu.UIMenuBuilder;
 import org.applied_geodesy.jag3d.ui.tabpane.UITabPaneBuilder;
 import org.applied_geodesy.jag3d.ui.tree.TreeItemValue;
 import org.applied_geodesy.jag3d.ui.tree.UITreeBuilder;
+import org.applied_geodesy.ui.dialog.OptionDialog;
 import org.applied_geodesy.util.ImageUtils;
 import org.applied_geodesy.util.i18.I18N;
 import org.applied_geodesy.util.sql.HSQLDB;
