+++ JAG3D/src/org/applied_geodesy/jag3d/ui/dialog/ApproximationValuesDialog.java
@@ -34,6 +34,7 @@ import org.applied_geodesy.jag3d.sql.PointTypeMismatchException;
 import org.applied_geodesy.jag3d.sql.SQLManager;
 import org.applied_geodesy.jag3d.sql.UnderDeterminedPointException;
 import org.applied_geodesy.jag3d.ui.tree.UITreeBuilder;
+import org.applied_geodesy.ui.dialog.OptionDialog;
 import org.applied_geodesy.util.i18.I18N;
 
 import javafx.application.Platform;
