+++ JAG3D/src/org/applied_geodesy/jag3d/ui/dialog/ColumnImportDialog.java
@@ -39,8 +39,8 @@ import org.applied_geodesy.jag3d.ui.io.CSVObservationFileReader;
 import org.applied_geodesy.jag3d.ui.io.CSVPointFileReader;
 import org.applied_geodesy.jag3d.ui.io.ColumnDefinedObservationFileReader;
 import org.applied_geodesy.jag3d.ui.io.ColumnDefinedPointFileReader;
-import org.applied_geodesy.jag3d.ui.textfield.LimitedTextField;
 import org.applied_geodesy.jag3d.ui.tree.TreeItemValue;
+import org.applied_geodesy.ui.textfield.LimitedTextField;
 import org.applied_geodesy.util.i18.I18N;
 import org.applied_geodesy.util.io.PreviewFileReader;
 import org.applied_geodesy.util.io.SourceFileReader;
