+++ JAG3D/src/org/applied_geodesy/jag3d/ui/dialog/FormatterOptionDialog.java
@@ -31,7 +31,8 @@ import java.util.Map;
 import java.util.Optional;
 
 import org.applied_geodesy.jag3d.sql.SQLManager;
-import org.applied_geodesy.jag3d.ui.table.CellValueType;
+import org.applied_geodesy.ui.dialog.OptionDialog;
+import org.applied_geodesy.util.CellValueType;
 import org.applied_geodesy.util.FormatterOptions;
 import org.applied_geodesy.util.FormatterOptions.FormatterOption;
 import org.applied_geodesy.util.i18.I18N;
