+++ JAG3D/src/org/applied_geodesy/jag3d/ui/dialog/RankDefectDialog.java
@@ -26,6 +26,7 @@ import java.util.Optional;
 import org.applied_geodesy.adjustment.network.DefectType;
 import org.applied_geodesy.adjustment.network.RankDefect;
 import org.applied_geodesy.jag3d.sql.SQLManager;
+import org.applied_geodesy.ui.dialog.OptionDialog;
 import org.applied_geodesy.util.i18.I18N;
 
 import javafx.application.Platform;
