+++ JAG3D/src/org/applied_geodesy/jag3d/ui/dialog/SearchAndReplaceDialog.java
@@ -25,6 +25,7 @@ import java.util.Optional;
 
 import org.applied_geodesy.jag3d.sql.SQLManager;
 import org.applied_geodesy.jag3d.ui.tree.TreeItemValue;
+import org.applied_geodesy.ui.dialog.OptionDialog;
 import org.applied_geodesy.util.ObservableLimitedList;
 import org.applied_geodesy.util.i18.I18N;
