+++ JAG3D/src/org/applied_geodesy/jag3d/ui/io/GSIFileReader.java
@@ -42,7 +42,6 @@ import org.applied_geodesy.jag3d.ui.tree.PointTreeItemValue;
 import org.applied_geodesy.jag3d.ui.tree.TreeItemType;
 import org.applied_geodesy.jag3d.ui.tree.TreeItemValue;
 import org.applied_geodesy.jag3d.ui.tree.UITreeBuilder;
-import org.applied_geodesy.util.io.DimensionType;
 import org.applied_geodesy.util.io.SourceFileReader;
 
 import javafx.scene.control.TreeItem;
