+++ JAG3D/src/org/applied_geodesy/jag3d/ui/io/xml/JobXMLFileReader.java
@@ -41,6 +41,7 @@ import org.applied_geodesy.adjustment.Constant;
 import org.applied_geodesy.adjustment.MathExtension;
 import org.applied_geodesy.adjustment.network.observation.reduction.Reduction;
 import org.applied_geodesy.jag3d.sql.SQLManager;
+import org.applied_geodesy.jag3d.ui.io.DimensionType;
 import org.applied_geodesy.jag3d.ui.table.row.PointRow;
 import org.applied_geodesy.jag3d.ui.table.row.TerrestrialObservationRow;
 import org.applied_geodesy.jag3d.ui.tree.ObservationTreeItemValue;
@@ -49,7 +50,6 @@ import org.applied_geodesy.jag3d.ui.tree.TreeItemType;
 import org.applied_geodesy.jag3d.ui.tree.TreeItemValue;
 import org.applied_geodesy.jag3d.ui.tree.UITreeBuilder;
 import org.applied_geodesy.util.XMLUtilities;
-import org.applied_geodesy.util.io.DimensionType;
 import org.applied_geodesy.util.io.SourceFileReader;
 import org.w3c.dom.Document;
 import org.w3c.dom.NamedNodeMap;
