+++ JAG3D/src/org/applied_geodesy/jag3d/ui/menu/UIMenuBuilder.java
@@ -54,28 +54,28 @@ import org.applied_geodesy.jag3d.sql.ProjectDatabaseStateType;
 import org.applied_geodesy.jag3d.sql.SQLManager;
 import org.applied_geodesy.jag3d.ui.JAG3D;
 import org.applied_geodesy.jag3d.ui.dialog.ColumnImportDialog;
-import org.applied_geodesy.jag3d.ui.dialog.OptionDialog;
 import org.applied_geodesy.jag3d.ui.io.BeoFileReader;
 import org.applied_geodesy.jag3d.ui.io.CongruenceAnalysisFlatFileReader;
 import org.applied_geodesy.jag3d.ui.io.DL100FileReader;
 import org.applied_geodesy.jag3d.ui.io.DefaultFileChooser;
+import org.applied_geodesy.jag3d.ui.io.DimensionType;
 import org.applied_geodesy.jag3d.ui.io.GSIFileReader;
 import org.applied_geodesy.jag3d.ui.io.M5FileReader;
 import org.applied_geodesy.jag3d.ui.io.ObservationFlatFileReader;
 import org.applied_geodesy.jag3d.ui.io.PointFlatFileReader;
 import org.applied_geodesy.jag3d.ui.io.ZFileReader;
+import org.applied_geodesy.jag3d.ui.io.report.FTLReport;
 import org.applied_geodesy.jag3d.ui.io.sql.OADBReader;
 import org.applied_geodesy.jag3d.ui.io.xml.HeXMLFileReader;
 import org.applied_geodesy.jag3d.ui.io.xml.JobXMLFileReader;
 import org.applied_geodesy.jag3d.ui.tree.TreeItemValue;
 import org.applied_geodesy.jag3d.ui.tree.UITreeBuilder;
+import org.applied_geodesy.ui.dialog.OptionDialog;
 import org.applied_geodesy.util.i18.I18N;
-import org.applied_geodesy.util.io.DimensionType;
 import org.applied_geodesy.util.io.LockFileReader;
 import org.applied_geodesy.util.io.SourceFileReader;
 import org.applied_geodesy.util.io.properties.HTTPPropertiesLoader;
 import org.applied_geodesy.util.io.properties.URLParameter;
-import org.applied_geodesy.util.io.report.FTLReport;
 import org.applied_geodesy.util.sql.DataBase;
 import org.applied_geodesy.util.sql.HSQLDB;
 import org.applied_geodesy.version.jag3d.DatabaseVersionMismatchException;
