+++ JAG3D/src/org/applied_geodesy/jag3d/ui/metadata/UIMetaDataPaneBuilder.java
@@ -25,9 +25,9 @@ import org.applied_geodesy.jag3d.sql.ProjectDatabaseStateChangedListener;
 import org.applied_geodesy.jag3d.sql.ProjectDatabaseStateEvent;
 import org.applied_geodesy.jag3d.sql.ProjectDatabaseStateType;
 import org.applied_geodesy.jag3d.sql.SQLManager;
-import org.applied_geodesy.jag3d.ui.dialog.OptionDialog;
-import org.applied_geodesy.jag3d.ui.textfield.LimitedTextArea;
-import org.applied_geodesy.jag3d.ui.textfield.LimitedTextField;
+import org.applied_geodesy.ui.dialog.OptionDialog;
+import org.applied_geodesy.ui.textfield.LimitedTextArea;
+import org.applied_geodesy.ui.textfield.LimitedTextField;
 import org.applied_geodesy.util.i18.I18N;
 
 import javafx.application.Platform;
