+++ JAG3D/src/org/applied_geodesy/jag3d/ui/table/ColumnTooltipHeader.java
@@ -21,6 +21,7 @@
 
 package org.applied_geodesy.jag3d.ui.table;
 
+import org.applied_geodesy.util.CellValueType;
 import org.applied_geodesy.util.FormatterChangedListener;
 import org.applied_geodesy.util.FormatterEvent;
 import org.applied_geodesy.util.FormatterEventType;
